move otaclient.app.errors as otaclient.errors
