+++ src/otaclient/boot_control/_grub.py
@@ -42,7 +42,7 @@ from pprint import pformat
 from subprocess import CalledProcessError
 from typing import ClassVar, Dict, Generator, List, Optional, Tuple
 
-from otaclient.app import errors as ota_errors
+from otaclient import errors as ota_errors
 from otaclient_api.v2 import types as api_types
 from otaclient_common._io import (
     read_str_from_file,
