+++ src/otaclient/boot_control/_jetson_uefi.py
@@ -32,7 +32,7 @@ from typing import Any, ClassVar, Generator, Literal
 from pydantic import BaseModel
 from typing_extensions import Self
 
-from otaclient.app import errors as ota_errors
+from otaclient import errors as ota_errors
 from otaclient.app.configs import config as cfg
 from otaclient.boot_control._firmware_package import (
     FirmwareManifest,
